Moved /template.js to /framework.template.js
